pilot/requirements.txt
google-adk>=1.16.0
fastapi
uvicorn
python-dotenv
websockets
httpx
pytest
pytest-mock
pytest-asyncio
cloud-sql-python-connector
pg8000
sqlalchemy
firebase-admin
opentelemetry-exporter-gcp-trace
opentelemetry-sdk
opentelemetry-api
motor